requirements.txt
streamlit==1.31.1
opencv-python==4.9.0.80
mediapipe==0.10.5
numpy==1.26.4
roboflow==1.1.19
pillow==10.2.0 